Revert "AGENDUM-62 DB Exceptions"

+++ user/dao.py
@@ -1,32 +1,14 @@
 from typing import List, Optional
 
-from mongoengine.errors import (
-    ValidationError,
-    NotUniqueError,
-)
+import sys
 
 from db_python_util.db_classes import User
 from db_python_util.db_helper import ConnectionManager
-from db_python_util.db_exceptions import (
-    EntityNotFoundException,
-    UsernameTakenException,
-)
 
 
 @ConnectionManager.requires_connection
-def get_by_id(id: str) -> User:
-    try:
-        user = User.objects.with_id(id)
-    except ValidationError: # An ObjectId "must be a 12-byte input or a 24-character hex string", but front-end shouldn't need to care about that
-        user = None
-
-    if user is None:
-        raise EntityNotFoundException(
-            User,
-            f"No user with id {id}"
-        )
-
-    return user
+def get_by_id(id: str) -> Optional[User]:
+    return User.objects.with_id(id)
 
 @ConnectionManager.requires_connection
 def get_by_username(name: str) -> Optional[User]:
@@ -48,29 +30,34 @@ def create(name: str, password_hash: str, password_salt: str) -> str:
         password_hash=password_hash,
         password_salt=password_salt
     )
-
-    try:
-        user.save()
-    except NotUniqueError:
-        raise UsernameTakenException(name)
-
+    user.save()
     return user.id.binary.hex()
 
 @ConnectionManager.requires_connection
-def update(id: str, name: str, password_hash: str):
-    user = get_by_id(id)
+def update(id: str, name: str, password_hash: str) -> bool:
+    user = User.objects.with_id(id)
+    if user is None:
+        sys.stderr.write(f"No user with id {id}!\n")
+        return False
 
     user_with_name = get_by_username(name)
     if (user_with_name is not None) and (user_with_name.id != user.id):
-        raise UsernameTakenException(name)
+        sys.stderr.write(f"The username '{name}' is already taken!\n")
+        return False
 
     user.update(
         username=name,
         display_name=name, # TODO: we need a display name
         password_hash=password_hash,
     )
+    return True
 
 @ConnectionManager.requires_connection
-def delete(id: str):
-    user = get_by_id(id)
+def delete(id: str) -> bool:
+    user = User.objects.with_id(id)
+    if id is None:
+        sys.stderr.write(f"No user with id {id}!\n")
+        return False
+
     user.delete()
+    return True
